feat: update docs and refactor feeMath
