+++ packages/dynamic-bonding-curve/src/math/feeMath.ts
@@ -1,11 +1,7 @@
 import BN from 'bn.js'
-import { pow, SafeMath } from './safeMath'
+import { SafeMath } from './safeMath'
 import { mulDiv } from './utilsMath'
-import {
-    BASIS_POINT_MAX,
-    FEE_DENOMINATOR,
-    MAX_FEE_NUMERATOR,
-} from '../constants'
+import { FEE_DENOMINATOR, MAX_FEE_NUMERATOR } from '../constants'
 import {
     BaseFeeMode,
     Rounding,
@@ -14,153 +10,11 @@ import {
     type PoolFeesConfig,
     type VolatilityTracker,
 } from '../types'
-
-/**
- * Get fee in period for exponential fee scheduler
- * @param cliffFeeNumerator Cliff fee numerator
- * @param reductionFactor Reduction factor
- * @param period Period
- * @returns Fee numerator
- */
-export function getFeeNumeratorInPeriod(
-    cliffFeeNumerator: BN,
-    reductionFactor: BN,
-    period: number
-): BN {
-    if (period === 0) {
-        return cliffFeeNumerator
-    }
-
-    if (period === 1) {
-        const basisPointMax = new BN(BASIS_POINT_MAX)
-        return mulDiv(
-            cliffFeeNumerator,
-            basisPointMax.sub(reductionFactor),
-            basisPointMax,
-            Rounding.Down
-        )
-    }
-
-    // calculate (1-reduction_factor/10_000)^period
-    const basisPointMax = new BN(BASIS_POINT_MAX)
-
-    // base = ONE_Q64 - (reductionFactor << RESOLUTION) / BASIS_POINT_MAX
-    const ONE_Q64 = new BN(1).shln(64)
-    const reductionFactorScaled = SafeMath.div(
-        SafeMath.shl(reductionFactor, 64),
-        basisPointMax
-    )
-    let base = SafeMath.sub(ONE_Q64, reductionFactorScaled)
-
-    const result = pow(base, new BN(period))
-
-    // final fee: cliffFeeNumerator * result >> 64
-    return SafeMath.div(SafeMath.mul(cliffFeeNumerator, result), ONE_Q64)
-}
-
-export function getFeeNumeratorFromAmount(
-    cliffFeeNumerator: BN,
-    referenceAmount: BN,
-    feeIncrementBps: BN,
-    inputAmount: BN
-): BN {
-    // if input amount is less than or equal to reference amount, return cliff fee
-    if (inputAmount.lte(referenceAmount)) {
-        return cliffFeeNumerator
-    }
-
-    // calculate fee increment numerator
-    const feeIncrementNumerator = mulDiv(
-        new BN(feeIncrementBps),
-        new BN(FEE_DENOMINATOR),
-        new BN(BASIS_POINT_MAX),
-        Rounding.Down
-    )
-
-    // calculate max index (how many increments until we reach MAX_FEE_NUMERATOR)
-    const deltaNumerator = new BN(MAX_FEE_NUMERATOR).sub(cliffFeeNumerator)
-    const maxIndex = deltaNumerator.div(feeIncrementNumerator)
-
-    // calculate a and b
-    // a = (input_amount - reference_amount) / reference_amount (integer division)
-    // b = (input_amount - reference_amount) % reference_amount (remainder)
-    const diff = inputAmount.sub(referenceAmount)
-    const a = diff.div(referenceAmount)
-    const b = diff.mod(referenceAmount)
-
-    // calculate fee numerator
-    let tradingFee: BN
-    if (a.lt(maxIndex)) {
-        // if a < max_index
-        // calculate fee using the formula:
-        // fee = x0 * (c + c*a + i*a*(a+1)/2) + b * (c + i*(a+1))
-        // where:
-        // x0 = reference_amount
-        // c = cliff_fee_numerator
-        // i = fee_increment_numerator
-        // a = (input_amount - x0) / x0 (integer division)
-        // b = (input_amount - x0) % x0 (remainder)
-
-        const one = new BN(1)
-        const two = new BN(2)
-
-        // c + c*a
-        const term1 = cliffFeeNumerator.add(cliffFeeNumerator.mul(a))
-
-        // i*a*(a+1)/2
-        const term2 = feeIncrementNumerator.mul(a).mul(a.add(one)).div(two)
-
-        // c + i*(a+1)
-        const term3 = cliffFeeNumerator.add(
-            feeIncrementNumerator.mul(a.add(one))
-        )
-
-        // calculate total fee
-        const totalFee = referenceAmount.mul(term1.add(term2)).add(b.mul(term3))
-
-        // convert to trading fee: trading_fee = total_fee / FEE_DENOMINATOR
-        tradingFee = totalFee.div(new BN(FEE_DENOMINATOR))
-    } else {
-        // if a >= max_index
-        // use MAX_FEE_NUMERATOR for the portion exceeding max_index
-        const one = new BN(1)
-        const two = new BN(2)
-
-        // c + c*max_index
-        const term1 = cliffFeeNumerator.add(cliffFeeNumerator.mul(maxIndex))
-
-        // i*max_index*(max_index+1)/2
-        const term2 = feeIncrementNumerator
-            .mul(maxIndex)
-            .mul(maxIndex.add(one))
-            .div(two)
-
-        // calculate fee for the first part (up to max_index)
-        const firstPartFee = referenceAmount.mul(term1.add(term2))
-
-        // calculate fee for the remaining part (beyond max_index)
-        const d = a.sub(maxIndex)
-        const leftAmount = d.mul(referenceAmount).add(b)
-        const secondPartFee = leftAmount.mul(new BN(MAX_FEE_NUMERATOR))
-
-        // calculate total fee
-        const totalFee = firstPartFee.add(secondPartFee)
-
-        // convert to trading fee: trading_fee = total_fee / FEE_DENOMINATOR
-        tradingFee = totalFee.div(new BN(FEE_DENOMINATOR))
-    }
-
-    // convert to fee numerator: fee_numerator = trading_fee * FEE_DENOMINATOR / input_amount
-    const feeNumerator = mulDiv(
-        tradingFee,
-        new BN(FEE_DENOMINATOR),
-        inputAmount,
-        Rounding.Up
-    )
-
-    // Cap at MAX_FEE_NUMERATOR
-    return BN.min(feeNumerator, new BN(MAX_FEE_NUMERATOR))
-}
+import {
+    getFeeNumeratorOnExponentialFeeScheduler,
+    getFeeNumeratorOnLinearFeeScheduler,
+} from './feeScheduler'
+import { getFeeNumeratorOnRateLimiter } from './rateLimiter'
 
 /**
  * Get current base fee numerator
@@ -170,7 +24,7 @@ export function getFeeNumeratorFromAmount(
  * @param inputAmount Input amount (optional, used for rate limiter)
  * @returns Current base fee numerator
  */
-export function getCurrentBaseFeeNumerator(
+export function getBaseFeeNumerator(
     baseFee: {
         cliffFeeNumerator: BN
         firstFactor: number
@@ -200,7 +54,7 @@ export function getCurrentBaseFeeNumerator(
             return baseFee.cliffFeeNumerator
         }
 
-        return getFeeNumeratorFromAmount(
+        return getFeeNumeratorOnRateLimiter(
             baseFee.cliffFeeNumerator,
             baseFee.thirdFactor, // reference amount
             new BN(baseFee.firstFactor), // fee increment bps
@@ -227,16 +81,14 @@ export function getCurrentBaseFeeNumerator(
 
         if (baseFeeMode === BaseFeeMode.FeeSchedulerLinear) {
             // linear fee calculation: cliffFeeNumerator - period * reductionFactor
-            const reduction = SafeMath.mul(new BN(period), baseFee.secondFactor)
-
-            if (reduction.gt(baseFee.cliffFeeNumerator)) {
-                return new BN(0)
-            }
-
-            return SafeMath.sub(baseFee.cliffFeeNumerator, reduction)
+            return getFeeNumeratorOnLinearFeeScheduler(
+                baseFee.cliffFeeNumerator,
+                baseFee.secondFactor,
+                period
+            )
         } else {
             // exponential fee calculation: cliff_fee_numerator * (1 - reduction_factor/10_000)^period
-            return getFeeNumeratorInPeriod(
+            return getFeeNumeratorOnExponentialFeeScheduler(
                 baseFee.cliffFeeNumerator,
                 baseFee.secondFactor,
                 period
@@ -299,7 +151,7 @@ export function getFeeOnAmount(
     volatilityTracker: VolatilityTracker
 ): FeeOnAmountResult {
     // get total trading fee
-    const baseFeeNumerator = getCurrentBaseFeeNumerator(
+    const baseFeeNumerator = getBaseFeeNumerator(
         poolFees.baseFee,
         currentPoint,
         activationPoint,
