+++ packages/dynamic-bonding-curve/tests/math/feeMath.test.ts
@@ -1,15 +1,12 @@
 import { test, expect } from 'bun:test'
-import {
-    getFeeNumeratorInPeriod,
-    getCurrentBaseFeeNumerator,
-    getVariableFee,
-} from '../../src/math/feeMath'
+import { getBaseFeeNumerator, getVariableFee } from '../../src/math/feeMath'
 import BN from 'bn.js'
 import { BaseFeeMode } from '../../src/types'
+import { getFeeNumeratorOnExponentialFeeScheduler } from '../../src/math/feeScheduler'
 
 test('getFeeInPeriod calculation', () => {
     // Test case 1: No reduction
-    const result1 = getFeeNumeratorInPeriod(
+    const result1 = getFeeNumeratorOnExponentialFeeScheduler(
         new BN(1000), // cliff fee
         new BN(0), // reduction factor
         0 // period as number, not BN
@@ -17,7 +14,7 @@ test('getFeeInPeriod calculation', () => {
     expect(result1.eq(new BN(1000))).toBe(true)
 
     // Test case 2: With reduction
-    const result2 = getFeeNumeratorInPeriod(
+    const result2 = getFeeNumeratorOnExponentialFeeScheduler(
         new BN(1000), // cliff fee
         new BN(100), // 1% reduction factor
         1 // period as number, not BN
@@ -28,7 +25,11 @@ test('getFeeInPeriod calculation', () => {
 
 test('getFeeInPeriod with higher periods', () => {
     // Test with period > 1 to test binary exponentiation
-    const result = getFeeNumeratorInPeriod(new BN(1000), new BN(100), 5)
+    const result = getFeeNumeratorOnExponentialFeeScheduler(
+        new BN(1000),
+        new BN(100),
+        5
+    )
 
     // Fee decreases with each period
     expect(result.lt(new BN(1000))).toBe(true)
@@ -36,7 +37,7 @@ test('getFeeInPeriod with higher periods', () => {
     expect(result.gte(new BN(0))).toBe(true)
 })
 
-test('getCurrentBaseFeeNumerator with linear mode', () => {
+test('getBaseFeeNumerator with linear mode', () => {
     const baseFee = {
         cliffFeeNumerator: new BN(1000),
         baseFeeMode: BaseFeeMode.FeeSchedulerLinear,
@@ -46,16 +47,12 @@ test('getCurrentBaseFeeNumerator with linear mode', () => {
     }
 
     // Before activation point
-    const result1 = getCurrentBaseFeeNumerator(baseFee, new BN(50), new BN(100))
+    const result1 = getBaseFeeNumerator(baseFee, new BN(50), new BN(100))
     // Use max period (min fee)
     expect(result1.eq(new BN(500))).toBe(true)
 
     // After activation point, 2 periods elapsed
-    const result2 = getCurrentBaseFeeNumerator(
-        baseFee,
-        new BN(300),
-        new BN(100)
-    )
+    const result2 = getBaseFeeNumerator(baseFee, new BN(300), new BN(100))
     expect(result2.eq(new BN(900))).toBe(true)
 })
 
@@ -69,7 +66,7 @@ test('getCurrentBaseFeeNumerator with exponential mode', () => {
     }
 
     // After activation point, 3 periods elapsed
-    const result = getCurrentBaseFeeNumerator(baseFee, new BN(350), new BN(100))
+    const result = getBaseFeeNumerator(baseFee, new BN(350), new BN(100))
 
     // Use exponential reduction
     expect(result.lt(new BN(1000))).toBe(true)
